adding get-outgoing-edges tool
